==> 6_gene_expression/12_annotate_with_symbols.ipynb <==
library(rtracklayer)
library(dplyr)

library(data.table)
library(stringr)

gtf <- rtracklayer::import("$PATH/gencode.v38.annotation.gtf")
gtf_df = as.data.frame(gtf)
print(dim(gtf_df))
print(head(gtf_df))

gtf_df$gene_no_ver <- str_split_fixed(gtf_df$gene_id, "\\.", 2)[,1]
print(head(gtf_df))

gtf_df <- gtf_df %>% select("gene_no_ver","gene_name")
gtf_df<- gtf_df[!duplicated(gtf_df), ]

head(gtf_df)

data <- as.data.frame(fread("$PATH/amppd_haplogroup_R_diff_exp_edgeR_results.csv"))
print(dim(data))
print(head(data))

data$gene_no_ver <- str_split_fixed(data$V1, "\\.", 2)[,1]
print(head(data))

#try inner join to see if any missing
print(dim(merge(x = data, y = gtf_df, by.x = 'gene_no_ver', by.y = 'gene_no_ver')))
new_data <- merge(x = data, y = gtf_df, by.x = 'gene_no_ver', by.y = 'gene_no_ver',all.x=TRUE)
print(dim(new_data))

new_data[is.na(new_data$symbol),]

#which ones are missing?
(data[!(data$gene_no_ver %in% gtf_df$gene_no_ver),])

dim(data[!(data$gene_no_ver %in% gtf_df$gene_no_ver),])

head(new_data)

new_data <- new_data[,c(1,13,3:12)]
new_data <- new_data[order(new_data$FDR),]
names(new_data)[names(new_data) == 'V1'] <- 'gene'
names(new_data)[names(new_data) == 'gene_name'] <- 'gene_symbol'
new_data$gene_symbol[is.na(new_data$gene_symbol)] <- " "
write.csv(new_data,"$PATH/amppd_haplogroup_R_diff_exp_edgeR_results_gene_symbol.csv",row.names=FALSE)


dim(new_data)

dim(data)

head(new_data)
